chore: Use utils for text
