+++ packages/ui/src/components/Overlay/ContextMenu/ContextMenu.tsx
@@ -2,13 +2,8 @@ import React, { forwardRef, ElementRef } from 'react'
 import { ComponentProps } from '@stitches/react'
 
 import { styled } from '~/theme'
-import { Text } from '~/components/Text'
-import { Switch } from '~/components/Input'
 
-/**
- * ContextMenu
- */
-const StyledContextMenu = styled('div', {
+const StyledBase = styled('div', {
   background: '$Background2dp',
   minWidth: '256px',
   width: '100%',
@@ -17,109 +12,17 @@ const StyledContextMenu = styled('div', {
   pointerEvents: 'all',
 })
 
-export interface ContextMenuProps
-  extends ComponentProps<typeof StyledContextMenu> {}
+export interface ContextMenuProps extends ComponentProps<typeof StyledBase> {}
 
 export const ContextMenu = forwardRef<
-  ElementRef<typeof StyledContextMenu>,
+  ElementRef<typeof StyledBase>,
   ContextMenuProps
 >((properties, forwardedRef) => {
   const { children, ...remainingProps } = properties
 
   return (
-    <StyledContextMenu ref={forwardedRef} {...remainingProps}>
+    <StyledBase ref={forwardedRef} {...remainingProps}>
       {children}
-    </StyledContextMenu>
-  )
-})
-
-/**
- * Item
- */
-const StyledItem = styled('div', {
-  display: 'flex',
-  height: '32px',
-  justifyContent: 'start',
-  alignItems: 'center',
-  paddingLeft: '16px',
-
-  '&:hover': {
-    background: '$ActionMain',
-  },
-})
-
-export interface ContextItemProps
-  extends ComponentProps<typeof StyledContextMenu> {}
-
-export const ContextItem = forwardRef<
-  ElementRef<typeof StyledContextMenu>,
-  ContextItemProps
->((properties, forwardedRef) => {
-  const { children, ...remainingProps } = properties
-
-  const isText = typeof children === 'string'
-
-  const ChildVariant = isText ? (
-    <Text color="Inherit">{children}</Text>
-  ) : (
-    children
-  )
-
-  return (
-    <StyledItem ref={forwardedRef} {...remainingProps}>
-      {ChildVariant}
-    </StyledItem>
-  )
-})
-
-/**
- * Toggle
- */
-const StyledToggle = styled('div', {
-  display: 'flex',
-  height: '32px',
-  width: '100%',
-  justifyContent: 'start',
-  alignItems: 'center',
-  paddingLeft: '16px',
-
-  '&:hover': {
-    background: '$ActionMain',
-  },
-})
-
-const SwitchContainer = styled('div', {
-  justifySelf: 'end',
-  marginLeft: 'auto',
-  display: 'flex',
-  justifyContent: 'end',
-  paddingRight: '12px',
-})
-
-export interface ContextToggleProps
-  extends ComponentProps<typeof StyledContextMenu> {}
-
-export const ContextToggle = forwardRef<
-  ElementRef<typeof StyledContextMenu>,
-  ContextToggleProps
->((properties, forwardedRef) => {
-  const { children, ...remainingProps } = properties
-
-  const isText = typeof children === 'string'
-
-  const ChildVariant = isText ? (
-    <Text color="Inherit">{children}</Text>
-  ) : (
-    children
-  )
-
-  return (
-    <StyledToggle ref={forwardedRef} {...remainingProps}>
-      {ChildVariant}
-
-      <SwitchContainer>
-        <Switch />
-      </SwitchContainer>
-    </StyledToggle>
+    </StyledBase>
   )
 })
