+++ packages/ui/src/components/Overlay/ContextMenu/index.ts
@@ -1,1 +1,3 @@
 export * from './ContextMenu'
+export * from './ContextToggle'
+export * from './ContextItem'
